migrate IAC files to separate dir
