+++ src/gatling/simulations/scenarios/iac/xuiIac.scala
@@ -1,9 +1,8 @@
-package scenarios
+package scenarios.iac
 
 import io.gatling.core.Predef._
 import io.gatling.http.Predef._
 import utils.{Common, Environment}
-import xui._
 
 object xuiIac {
 
@@ -12,7 +11,7 @@ object xuiIac {
   val Login =
 
     exec(XuiHelper.Homepage)
-    .exec(XuiHelper.Login("#{user}", "#{password}"))
+    .exec(XuiHelper.Login("#{email}", "#{password}"))
 
   val Logout =
